added subject 1 data, non processed
I added subject 1 data from the DB2 set on Ninapro. To get relevant data for config of the PGA, you need to import "dataForPGA_khubbard.mat" and this will give you two variables. The relevant features for configuration are located in "pgaConfigData" with the following interpretation:

Row 1: corresponds to the RMS value of each channel

Row 2: the maximum value of the channel

Row 3: the minimum value of the channel

and most importantly
Row 4: The peak to peak value of the channel

All channel placements are exactly as discussed in the "DATA_ACQUISITION_PROCEDURE.docx". I have already mapped the channels accordingly so you can just focus on "Channel 1,2,3,..." etc.